backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==1.4.46
psycopg2-binary==2.9.6
pydantic==1.10.7
python-multipart==0.0.5
